Venue events modal now is has content loaded from venue_events.html.
The content is rendered by the server prior to being displayed in the
modal.

All that is missing is the reading of a given venue's events and routing
that info back to the client through venue_events.html... but that should
be trivial once I have the files all set up.